pythonFunctions/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
async-timeout==4.0.3
attrs==23.2.0
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
configobj==5.0.8
configparser==7.0.0
crewai==0.5.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
exceptiongroup==1.2.2
filelock==3.15.4
fitz==0.0.1.dev2
Flask==3.0.3
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.0
idna==3.7
isodate==0.6.1
itsdangerous==2.2.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.1.0
langchain-community==0.0.20
langchain-core==0.1.23
langchain-openai==0.0.2.post1
langsmith==0.0.87
looseversion==1.3.0
lxml==5.2.2
MarkupSafe==2.1.5
marshmallow==3.21.3
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.2.1
nibabel==5.2.1
nipype==1.8.6
numpy==1.26.4
openai==1.37.0
packaging==23.2
pandas==2.2.2
pathlib==1.0.1
prov==2.0.1
pydantic==2.8.2
pydantic_core==2.20.1
pydot==3.0.1
pyparsing==3.1.2
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
pytz==2024.1
pyxnat==1.6.2
PyYAML==6.0.1
rdflib==6.3.2
regex==2024.7.24
requests==2.32.3
scipy==1.13.1
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.31
tenacity==8.5.0
tiktoken==0.5.2
tqdm==4.66.4
traits==6.3.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4

python-dotenv~=1.0.1